Move build-scripts/build-all-guides.py to ssg/_build_guides.py
